Исправил замечания наставника по прошлому модулю
