+++ src/utils.js
@@ -18,7 +18,7 @@ const createElement = (template) => {
   const newElement = document.createElement(`div`);
   newElement.innerHTML = template;
 
-  return newElement.firstChild;
+  return newElement.firstElementChild;
 };
 
 const render = (container, element, place) => {
